renamte toast.js to sf-toast.js
