+++ src/cards/lights/card.js
@@ -3,7 +3,7 @@ import {unsafeHTML} from 'lit/directives/unsafe-html.js';
 import {isEqual} from 'lodash-es';
 
 import '../../components/sf-tiles.js';
-import '../../components/toast.js';
+import '../../components/sf-toast.js';
 import {House} from '../../helpers/entities/house.js';
 import {
   ENTITY_KIND_LIGHT,
